+++ flutter/android/src/main/kotlin/io/sentry/flutter/SentryFlutterPlugin.kt
@@ -29,14 +29,13 @@ import java.lang.ref.WeakReference
 import java.util.Locale
 import java.util.UUID
 
-@Suppress("TooManyFunctions")
 class SentryFlutterPlugin : FlutterPlugin, MethodCallHandler, ActivityAware {
   private lateinit var channel: MethodChannel
   private lateinit var context: Context
 
   private var activity: WeakReference<Activity>? = null
   private var framesTracker: ActivityFramesTracker? = null
-  private var autoPerformanceTrackingEnabled = false
+  private var autoPerformanceTracingEnabled = false
 
   override fun onAttachedToEngine(flutterPluginBinding: FlutterPlugin.FlutterPluginBinding) {
     context = flutterPluginBinding.applicationContext
@@ -155,14 +154,24 @@ class SentryFlutterPlugin : FlutterPlugin, MethodCallHandler, ActivityAware {
         // options.isEnableNdk = false
       }
 
-      args.getIfNotNull<Boolean>("enableAutoPerformanceTracking") { enableAutoPerformanceTracking ->
-        if (enableAutoPerformanceTracking) {
-          autoPerformanceTrackingEnabled = true
+      args.getIfNotNull<Boolean>("enableAutoPerformanceTracing") { enableAutoPerformanceTracing ->
+        if (enableAutoPerformanceTracing) {
+          autoPerformanceTracingEnabled = true
           framesTracker = ActivityFramesTracker(LoadClass(), options)
         }
       }
 
       args.getIfNotNull<Boolean>("sendClientReports") { options.isSendClientReports = it }
+
+      args.getIfNotNull<Map<String, Any>>("sdk") { sdk ->
+        val name = sdk["name"] as? String
+        val version = sdk["version"] as? String
+        if (name != null && version != null) {
+          val sdkVersion = SdkVersion(name, version)
+          options.setSentryClientName(name)
+          options.setSdkVersion(sdkVersion)
+        }
+      }
 
       options.setBeforeSend { event, _ ->
         setEventOriginTag(event)
@@ -170,13 +179,13 @@ class SentryFlutterPlugin : FlutterPlugin, MethodCallHandler, ActivityAware {
         event
       }
 
-      // missing proxy, enableScopeSync
+      // missing proxy
     }
     result.success("")
   }
 
   private fun fetchNativeAppStart(result: Result) {
-    if (!autoPerformanceTrackingEnabled) {
+    if (!autoPerformanceTracingEnabled) {
       result.success(null)
       return
     }
@@ -199,7 +208,7 @@ class SentryFlutterPlugin : FlutterPlugin, MethodCallHandler, ActivityAware {
   }
 
   private fun beginNativeFrames(result: Result) {
-    if (!autoPerformanceTrackingEnabled) {
+    if (!autoPerformanceTracingEnabled) {
       result.success(null)
       return
     }
@@ -212,7 +221,7 @@ class SentryFlutterPlugin : FlutterPlugin, MethodCallHandler, ActivityAware {
 
   private fun endNativeFrames(id: String?, result: Result) {
     val activity = activity?.get()
-    if (!autoPerformanceTrackingEnabled || activity == null || id == null) {
+    if (!autoPerformanceTracingEnabled || activity == null || id == null) {
       if (id == null) {
         Log.w("Sentry", "Parameter id cannot be null when calling endNativeFrames.")
       }
@@ -263,7 +272,6 @@ class SentryFlutterPlugin : FlutterPlugin, MethodCallHandler, ActivityAware {
     }
   }
 
-  @Suppress("ComplexMethod")
   private fun setUser(user: Map<String, Any?>?, result: Result) {
     if (user == null) {
       Sentry.setUser(null)