+++ CHANGELOG.md
@@ -5,6 +5,16 @@
 ### Features
 
 - Error Cause Extractor ([#1198](https://github.com/getsentry/sentry-dart/pull/1198))
+### Breaking Changes
+
+- Rename APM tracking feature flags to tracing ([#1222](https://github.com/getsentry/sentry-dart/pull/1222))
+  - Rename
+    - enableAutoPerformanceTracking to enableAutoPerformanceTracing
+    - enableOutOfMemoryTracking to enableWatchdogTerminationTracking
+
+### Enhancements
+
+- Migrate to sentry cocoa v8 ([#1197](https://github.com/getsentry/sentry-dart/pull/1197))
 
 ## 7.0.0-alpha.2
 
@@ -15,9 +25,10 @@
 ### Breaking Changes
 
 - Enable APM features by default ([#1217](https://github.com/getsentry/sentry-dart/pull/1217))
-  - captureFailedRequests
-  - enableStructuredDataTracing
-  - enableUserInteractionTracing
+  - Enable by default
+    - captureFailedRequests
+    - enableStructuredDataTracing
+    - enableUserInteractionTracing
 - Mark transaction as internal_error in case of unhandled errors ([#1218](https://github.com/getsentry/sentry-dart/pull/1218))
 
 ## 7.0.0-alpha.1